update DA14580 SDK to 504

+++ PrjEmcWxDa1458x/sdk_580/sdk/platform/driver/uart/uart.h
@@ -168,7 +168,7 @@ void uart_init(uint8_t baudr, uint8_t mode);
  * @return void
  *****************************************************************************************
  */
-void uart2_init(uint8_t baudr, uint8_t mode);
+void uart2_init(uint16_t baudr, uint8_t mode);
 
 /**
  ****************************************************************************************
